linking to CAMediaTiming Protocol instead of grand central dispatch, more polish

+++ Loading/LoadingView.h
@@ -5,7 +5,7 @@
 @interface LoadingView : UIImageView
 
 @property (nonatomic, strong) UIColor *color;
-@property float speed;
+@property CGFloat speed;
 @property BOOL isAnimating;
 
 - (id)initWithCircleSize:(float)size;
